% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{calculate_celltype_prob}
\alias{calculate_celltype_prob}
\title{Calculate the cell type probablity assignment according to a markers database}
\usage{
calculate_celltype_prob(clt_marker_list, marker_database_list, type = "jacc")
}
\arguments{
\item{clt_marker_list}{This is the result returned from the function 'curate_markers', a list object containing marker genes for each specific cluster.}

\item{type}{This determines the type of overlap to calculate. Defaults to the Jaccard index, "jacc". Accuracy ("ac") and F1 ("f1") are also available. We recommend using jaccard or accuracy in applications.}

\item{marker_database}{A list object that contains the marker genes from each specific cell type. A pre-processed variable is stored as 'panglao_marker_list', which contains markers for neural and immune cell types.}
}
\description{
This function calculates probability of each cell type in the desired database, according to the number of overlapping genes. We have included the "Panglao" database as of date 7th Feb 2020 in our package.
}
\section{Biological Analysis}{

}

\examples{
gene_expression_data <- read.csv("single_cell_dataset.csv", header = T, row.names = 1) # make sure that genes are in rows
cluster_vector <- read.table("results_from_rpkmeans.txt")$V1
eval_gene_markers <- store_markers(gene_expression_data, cluster_vector, threads = 8)
gene_names = rownames(gene_expression_data)
cluster_markers_list <- curate_markers(eval_gene_markers, gene_names, wilcox_threshold=0.001, logfc_threshold=1.5)

celltype_prob <- calculate_celltype_prob(cluster_markers_list, panglao_marker_list, type = "jacc")
celltype_id <- rowMaxs(celltype_prob)
}
\keyword{calculate_celltype_prob}
