% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{calculate_enrichment_prob}
\alias{calculate_enrichment_prob}
\title{Calculate the enrichment probablity of each cell type based on a hypergeometric distribution.}
\usage{
calculate_enrichment_prob(
  clt_marker_list,
  marker_databse_list,
  total_background_genes = 10000,
  type = "jacc"
)
}
\arguments{
\item{clt_marker_list}{This is the result returned from the function 'curate_markers', a list object containing marker genes for each specific cluster.}

\item{total_background_genes}{This is the number of genes in your dataset}

\item{type}{This determines the type of overlap to calculate. Defaults to the Jaccard index, "jacc". Accuracy ("ac") and F1 ("f1") are also available. We recommend using jaccard or accuracy in applications.}

\item{marker_database}{A list object that contains the marker genes from each specific cell type. A pre-processed variable is stored as 'panglao_marker_list', which contains markers for neural and immune cell types.}
}
\description{
This function calculates enrichment probability for each cell type vs each cluster. In principle, it determines the statistical validity of the cell type assignment, based on the number of overlapping genes.
}
\section{Biological Analysis}{

}

\examples{
gene_expression_data <- read.csv("single_cell_dataset.csv", header = T, row.names = 1) # make sure that genes are in rows
cluster_vector <- read.table("results_from_rpkmeans.txt")$V1
eval_gene_markers <- store_markers(gene_expression_data, cluster_vector, threads = 8)
gene_names = rownames(gene_expression_data)
cluster_markers_list <- curate_markers(eval_gene_markers, gene_names, wilcox_threshold=0.001, logfc_threshold=1.5)

n_genes <- nrow(gene_expression_data)
enrichment_prob <- calculate_enrichment_prob(cluster_markers_list, panglao_marker_list, n_genes, type = "jacc")
}
\keyword{calculate_enrichment_prob}
