% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{curate_markers}
\alias{curate_markers}
\title{Find the expression markers specific to your cluster assignments}
\usage{
curate_markers(
  whole_list,
  gene_names,
  wilcox_threshold = 0.001,
  logfc_threshold = 1.5
)
}
\arguments{
\item{whole_list}{This is the result returned from the function 'store_markers', containing the wilcox test and log fold change values.}

\item{gene_names}{A character vector containing the gene names in the same order as your input gene expression matrix. e.g. c("SOX2", "ALDOC")}

\item{wilcox_threshold}{This is the p-value cut-off for the wilcox test, it is set to 0.001 by default. Recommend a smaller value for a stricter detection}

\item{logfc_threshold}{This is the log fold change threshold, which is 1.5 by default. This can be increased for stricter restriction on sample sizes.}
}
\description{
This function calculates the wilcox rank sum test and log fold change ratio for all genes in each cluster.
}
\section{Biological Analysis}{

}

\examples{
gene_expression_data <- read.csv("single_cell_dataset.csv", header = T, row.names = 1) # make sure that genes are in rows
cluster_vector <- read.table("results_from_rpkmeans.txt")$V1
eval_gene_markers <- store_markers(gene_expression_data, cluster_vector, threads = 8)
gene_names = rownames(gene_expression_data)
cluster_markers_list <- curate_markers(eval_gene_markers, gene_names, wilcox_threshold=0.001, logfc_threshold=1.5)
}
\keyword{curate_markers}
