% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{find_specific_marker}
\alias{find_specific_marker}
\title{Utility function for calculating the overlaps of specific markers vs each cluster}
\usage{
find_specific_marker(gene_name, f_list, type = "jacc")
}
\arguments{
\item{gene_name}{List of gene names that we want to query in the clusters.}

\item{f_list}{A list object that contains the marker genes from each specific cluster.}

\item{type}{This determines the type of overlap to calculate. Defaults to the Jaccard index, "jacc". Accuracy ("ac") and F1 ("f1") are also available. We recommend using jaccard or accuracy in applications.}
}
\description{
This is the utility function behind calculate_celltype_prob, which implements the Jaccard, accuracy and F1 to calculate the value of overlaps between marker genes and cluster specific genes.
}
\section{Biological Analysis}{

}

\keyword{find_specific_marker}
