% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_marker_expressions}
\alias{plot_marker_expressions}
\title{Plot for visualizing cluster-specific markers}
\usage{
plot_marker_expressions(processed_list, gene_levels = NULL, cell_levels = NULL)
}
\arguments{
\item{processed_list}{This is the pre-processed list from the function 'process_marker_expression', which returns an average expression matrix and the percentage matrix (ratio of zeros)}

\item{gene_levels}{This is a character vector containing gene names in the order that you want to display in the plot}

\item{cell_levels}{This is a character vector containing cell type annotations / cluster labels in the order that you want to display in the plot}
}
\description{
This function plots the average expression values (denoted by color of dots) and the percentage of zeros in the cluster (denoted size of dots)
}
\section{Biological Analysis}{

}

\examples{
gene_expression_data <- read.csv("single_cell_dataset.csv", header = T, row.names = 1) # make sure that genes are in rows
cluster_vector <- read.table("results_from_rpkmeans.txt")$V1

# example marker list:
selected_marker_genes <- c("SOX2", "ALDOC", "CCND2", "OLIG1", "OLIG2")
gene_expression_subset <- gene_expression_data[match(tolower(selected_marker_genes), tolower(rownmaes(gene_expression_data))), ]
processed_markers <- process_marker_expression(gene_expression_subset, cluster_vector)
cell_levels <- unique(cluster_vector)
gene_levels <- selected_marker_genes
marker_plot <- plot_marker_expression(processed_markers, gene_levels=gene_levels, cell_levels=cell_levels)
}
\keyword{plot_marker_expressions}
