% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{process_marker_expression}
\alias{process_marker_expression}
\title{Prepares the data for plotting marker genes}
\usage{
process_marker_expression(gene_cell_matrix, cell_labels)
}
\arguments{
\item{gene_cell_matrix}{This is the pre-processed gene expression matrix (normalized and logged). The subset expression from the markers list is used}

\item{cell_labels}{Cell labels which can be annotated or just cluster labels.}
}
\description{
This function calculates the cluster-specific average gene expressions and the percentage of zeros in each cluster. Prepares the data for plotting
}
\section{Biological Analysis}{

}

\examples{
gene_expression_data <- read.csv("single_cell_dataset.csv", header = T, row.names = 1) # make sure that genes are in rows
cluster_vector <- read.table("results_from_rpkmeans.txt")$V1

# example marker list:
selected_marker_genes <- c("SOX2", "ALDOC", "CCND2", "OLIG1", "OLIG2")
gene_expression_subset <- gene_expression_data[match(tolower(selected_marker_genes), tolower(rownmaes(gene_expression_data))), ]
processed_markers <- process_marker_expression(gene_expression_subset, cluster_vector)
}
\keyword{process_marker_expression}
