% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{process_panglao_db}
\alias{process_panglao_db}
\title{Pre-processes the Panglao database to curate markers for desired cell types.}
\usage{
process_panglao_db(cell_type_name_list, database)
}
\arguments{
\item{cell_type_name_list}{This is the corresponding cell type names in the database.}

\item{database}{The database table which is loaded.}
}
\description{
This function processes the panglao database to the format of lists, where each list is a cell type containing marker genes.
}
\section{Biological Analysis}{

}

\examples{
panglao_table <-read.csv("panglao_database.csv", header = T, sep = "\t")
type_names <- c("astrocytes", "oligodendrocytes")
preprocessed_pl <- process_panglao_db(type_names, panglao_table)

# Preloaded Panglao database objects:
# panglao_db: the full panglao marker list as of date 7th Feb 2020
# panglao_marker_list: the pre-processed list for all neural and immune cell types.
}
\keyword{calculate_enrichment_prob}
