% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{store_markers}
\alias{store_markers}
\title{Calculate the wilcox rank sum test and log fold change values}
\usage{
store_markers(input_data_matrix, labels_vector, threads = 8)
}
\arguments{
\item{input_data_matrix}{A pre-processed gene expression matrix (ie normalization and log). Matrix input is a genes (row) by cells (col) gene expression matrix with all genes.}

\item{labels_vector}{A numeric vector containing the cell cluster assignment of each cell. If minimum value is 0, it will automatically add 1 to the vector.}

\item{threads}{The number of threads used for parallel computing. Defaults to 8}
}
\description{
This function calculates the wilcox rank sum test and log fold change ratio for all genes in each cluster.
}
\section{Biological Analysis}{

}

\examples{
gene_expression_data <- read.csv("single_cell_dataset.csv", header = T, row.names = 1) # make sure that genes are in rows
cluster_vector <- read.table("results_from_rpkmeans.txt")$V1
eval_gene_markers <- store_markers(gene_expression_data, cluster_vector, threads = 8)
}
\keyword{store_markers}
